move postkeys function to keystore package
